# server/swagger.yaml
swagger: "2.0"
info:
  description: "Todo app API"
  version: "1.0.0"
  title: "Todo"
host: "localhost:5000"
schemes:
- "http"
paths:
  /todo_lists:
    get:
      summary: "Get list of all todo lists"
      operationId: "getTodoLists"
      produces:
      - "application/json"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TodoList"

    post:
      summary: "Create a new todo list"
      operationId: "createTodoList"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NewTodoList"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/TodoList"

  /todo_lists/{listId}:
    get:
      summary: "Find todo list by ID"
      operationId: "getTodoListById"
      produces:
      - "application/json"
      parameters:
      - name: "listId"
        in: "path"
        description: "ID of todo list to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TodoList"
        404:
          description: "Todo list not found"
    delete:
      summary: "Delete todo list by ID"
      operationId: "deleteTodoListById"
      parameters:
      - name: "listId"
        in: "path"
        description: "ID of todo list to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Success"
        404:
          description: "Todo list not found"
    patch:
      summary: "Update todo list by ID"
      operationId: "updateTodoListById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "listId"
        in: "path"
        description: "ID of todo list to update"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/TodoListUpdate"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TodoList"
        404:
          description: "Todo list not found"
  /todo:
    post:
      summary: "Create a new todo item"
      operationId: "createTodo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NewTodo"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Todo"
  /todos:
    get:
      summary: "Show all todos"
      operationId: "getTodos"
      produces:
      - "application/json"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Todo"
        404:
          description: "Todos not found"
    post:
      summary: "Find todos by title"
      operationId: "getTodosByTitle"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Todo"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Todo"
        404:
          description: "Todos not found"
  /todo/{todoId}:
    get:
      summary: "Find todo by ID"
      operationId: "getTodoById"
      produces:
      - "application/json"
      parameters:
      - name: "todoId"
        in: "path"
        description: "ID of todo to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Todo"
        404:
          description: "Todo not found"
    patch:
      summary: "Update todo by ID"
      operationId: "updateTodoById"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
      - name: "todoId"
        in: "path"
        description: "ID of todo to update"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/TodoUpdate"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Todo"
        404:
          description: "Todo not found"
    delete:
      summary: "Delete todo by ID"
      operationId: "deleteTodoById"
      parameters:
      - name: "todoId"
        in: "path"
        description: "ID of todo to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        204:
          description: "Success"
        404:
          description: "Todo not found"
  
definitions:
  NewTodoList:
    type: "object"
    properties:
      name:
        type: "string"
  TodoList:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      todos:
        type: "array"
        items:
          type: "integer"
          format: "int64"
  NewTodo:
    type: "object"
    properties:
      title:
        type: "string"
      todo_list_id:
        type: "integer"
        format: "int64"

  Todo:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      todo_list_id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      completed:
        type: "boolean"
      created:
        type: "string"
      modified:
        type: "string"
  TodoUpdate:
    type: "object"
    properties:
      title:
        type: "string"
      completed:
        type: "boolean"

        